+++ src/index.html
@@ -16,6 +16,7 @@
 
   </section>
   <footer>Copyright</footer>
+  <script type="module" src="main.js"></script>
   <script type="module" src="tempString.js"></script>
   <script src="https://www.gstatic.com/firebasejs/5.8.1/firebase.js"></script>
   <script src="https://www.gstatic.com/firebasejs/5.5.2/firebase-app.js"></script>
@@ -23,6 +24,6 @@
   <script type="module" src="config.js"></script>
   <script type="module" src="lib/index.js"></script>
   <script type="module" src="app.js"></script>
-  <script type="module" src="main.js"></script>
+
 </body>
 </html>
